Removed Renderer trait and added generic  method to Form

+++ src/utils.rs
@@ -2,16 +2,16 @@
 //! Miscellaneous utility functions.
 //!
 
-use std::f64::consts::PI;
+use std::f32::consts::PI;
 use std::num::{Float, Int};
 
 /// Convert turns to radians.
-pub fn turns(t: f64) -> f64 {
+pub fn turns(t: f32) -> f32 {
     2.0 * PI / t
 }
 
 /// Convert degrees to radians.
-pub fn degrees(d: f64) -> f64 {
+pub fn degrees(d: f32) -> f32 {
     d * PI / 180.0
 }
 
@@ -26,18 +26,18 @@ pub fn modulo<I: Int>(a: I, b: I) -> I {
 }
 
 /// Modulo float.
-pub fn fmod(f: f64, n: i64) -> f64 {
-    let i = f.floor() as i64;
-    modulo(i, n) as f64 + f - i as f64
+pub fn fmod(f: f32, n: i32) -> f32 {
+    let i = f.floor() as i32;
+    modulo(i, n) as f32 + f - i as f32
 }
 
 /// Return the max between to floats.
-pub fn max(a: f64, b: f64) -> f64 {
+pub fn max(a: f32, b: f32) -> f32 {
     if a >= b { a } else { b }
 }
 
 /// Return the min between to floats.
-pub fn min(a: f64, b: f64) -> f64 {
+pub fn min(a: f32, b: f32) -> f32 {
     if a <= b { a } else { b }
 }
